Update implementation:
Fix Dapper repository, add EF repository, kafka sent notify messages service implementation, impreve api project, fix swagger documentation, add db generator project, add client generator project, add docker-compose environment config from postgresql, kafka, add custom auth middleware, improve api versioning, add api controller for api token menedgement, fix kestrel config, add correlation ids, add healthcheck, heartbeat metrics.

+++ src/Orange.ApiTokenValidation.Domain/Models/TokenValidationResult.cs
@@ -3,8 +3,8 @@
     public class TokenValidationResult
     {
         /// <summary> 
-        /// Expiration
+        /// Expiration in seconds
         /// </summary>
-        public long Expiration { get; set; }
+        public double Expiration { get; set; }
     }
 }
